requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1;python_version<"3.9"
black==22.12.0
click==8.1.3
colorama==0.4.5
cssbeautifier==1.14.6
Django==3.2.16
django-dirtyfields==1.9.1
django-environ==0.9.0
django-extensions==3.2.1
djangorestframework==3.14.0
djlint==1.12.3
EditorConfig==0.12.3
factory-boy==3.2.1
Faker==15.3.3
flake8==4.0.1
html-tag-names==0.1.2
html-void-elements==0.1.0
importlib-metadata==4.12.0
jsbeautifier==1.14.6
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.3.0
platformdirs==2.6.0
psycopg==3.1.4
psycopg2-binary==2.9.5
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
regex==2022.8.17
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
typed-ast==1.5.4
typing_extensions==4.2.0
zipp==3.8.0
